requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0
anthropic>=0.39
openai>=1.50
